Fixed number of bugs that were causing unit test failures.  2 failures exist still, but I'll let someone else know them out since they are in the munitions stuff and look complex to track down.
All of the CEGUI stuff now has CEGUI/ in front of the path so that the sub directory in the project doesn't have to be added to the path.  

The Stealth viewer starts up now, but can't really be used because no map exists and the models haven't been moved over.