Fixed an obscure bug, apparently a physics simulation has to be simulated after any one object is made, before you can call a 'get' on the actor. The get will not return you the correct information until after the system has stepped once.
This is not a fix for ageia in general but more of a fix for us and note for future problems that might come up for using gets/sets on newly made objects. To work around this one frame, if the time is 0 we just dont update the object, however this is not really correct, and theres no other good way around it. 

We would need to go ahead and put ageia in a different area for updating. IE On framesync method or another place / thread that doesnt really go with the components / actors.