Changed the HLA window so that it will not enable the disconnect button if the application is not currently connected.
The HLA Connection Component will only perform the disconnect procedure if the application is in a connected state. It also determines if the specified map is valid before attempting a connection to the HLA network. The old code did not catch an exception that was being thrown or would just set the state to connected when it really was not.

This seems to fix the connection parameters issue. The issue stated that a crash would happen when closing the application after switching to a second connection and deleting the first; DVTE bug# 699.