cleaned up the base inpuc code quite a bit, which affected the driver demo quite a bit. Removed uses of the SetAgeioUserData everywhere because using it is no longer necessary, and really shouldn't be set to anything but the helper.
Moved the ephemeris and fog sphere to 999 instead of 9000.  Since they both write 1 to the depth buffer, the spheres just have to exist between near and far so they won't be clipped. Having them at 9000 used to be necessary because the ephemeris didn't move in the z direction, but now that has been changed.  Having it 999 means the far can be anything 1000 or greater, and the near can be anything up to 999.  Those seemed like reasonable values.

Fixed a few unit test errors, although one reemains.

Also made the code use a method SetGroupCollision on the physics component instead of the direct PhysX call because it will make more sense when using dtPhysics.