Created a place for the AudioStreamer class.  It's primary purpose is to provide support for compressed audio formats.  FFMPEG is the key library to all for this.
This class was not added to Delta3D proper because it was deemed that the dependencies were too ugly.  Compiling FFMPEG on Windows requires MinGW and a slightly modification to Visual Studio 2008/2005 headers to make it C99 compliant (see docs on FFMPEG's website).

OTOH, the dependencies are relatively easy to handle on package-based flavors of Linux (Fedora, Ubuntu, Debian, etc).   Also, it is possible that a future version of Visual Studio will be C99 compliant.  Then FFMPEG would be able to provide precompiled libs and binaries for Windows -- or at least make the resolving the dependencies less ugly.