Added a new DRPublishingActComp to SimCore - pulled hundreds of lines of code out of BaseEntity and BasePhysicsVehicle. This goes in line with a major refactor of the way the actor components are initialized in Delta.
To sum up, there is a BuildActorComponents() method on GameActor that is called during the initialization of a GameActorProxy. This allows you to setup your actor components and initialize them without having to put them inside of BuildPropertyMaps(). 

The major purpose of all this was to allow initialization of actor components.