Moved GetMPH up to the BaseWheeledVehiclePhysicsHelper,  made the code only give the MPH based on the actual travel direction so it doesn't give a result when sliding to the side.
It also gives a -MPH on the physics helper for the sake of speed limiting. 

The FourWheelVehicleActor now overrides GetMPH to return the one from the physics helper, but it returns the ABS.  

Added const to GetMPH and GetHasDriver everywhere including on the vehicle interface, and made some other methods const to handle it properly.