A heavy rewrite of the whole dtScript system. This was necessary because my initial try at a generic template based wrapper manager was basically shit. The Wrap() and Unwrap() functions used a lot of template magic to select the correct wrapper. If the include file of the wrapper was not included before wrap or unwrap were called, the compile did not fail but also did not do anything useful. So there was always a lot of wondering why something could not be wrapped or why some variable did not show up in the js context.
Now each wrapper has to provide a WrapVec3 and UnwrapVec3 function (of
course with Vec3 replaced with the thing to wrap). And they have to be
called that way.

This release throws out all the wrappers of dtGame actors, components
etc. I may re-include them at some point but for now they are out.
There is probably no-one using this stuff that is not using dtEntity.
If you are, drop me a forum message and I will re-include the dtGame
wrappers.

Also did some code style updates.