Changed the human so that it doesn't have a motionless state.  The Walk-Run blend supports being still now, so it's not necessary and simplifies the planner.
This also means the crawl and low walk blends have a blending and speed scaling, but they use the walk speed as the inherent at the moment, which is clearly wrong, though there was no such thing before, so it's not really worse.  The code needs to figure out the inherent speed of these things, which is the source of not wanting to add properties for all of these things.

A lot of the constants are commented out for now.  I will clean this out later.

The human tests have been updated to reflect the changes.  
This also includes some fixes for deprecation warnings.